adding `return_parameters` option to Engine
Adding optional parameter `return_parameters` which returns each the parameters used in each run within the experiment results output.

Defaults to False.

This feature saves time, reduces complexity, and avoids the risk of error when requiring users to manually construct a map of which parameters correspond to each run after the fact.